# specs/001-supervised-autonomous-loop/contracts/http-api.openapi.yaml
openapi: 3.1.0
...
info:
  title: Moonshift Supervised Autonomous Loop API
  version: 1.0.0-planned
  description: >-
    Loopback-only Iteration 0 contract for the single supervisor. This is a planning artifact,
    not a shipped or supported API.
servers:
  - url: http://127.0.0.1:{port}
    description: Local evaluation control plane
    variables:
      port:
        default: "4310"
security:
  - LocalSupervisorSession: []
...
paths:
...
  /v1/projects/{projectId}/approvals:
    get:
      tags: [Approvals]
      operationId: listApprovals
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - name: state
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/ApprovalState"
      responses:
        "200":
          description: >-
            Approval projections visible to the supervisor. Each item exposes its aggregate version;
            retrieve the selected item to obtain the strong ETag required for a decision.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupervisionView"
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
  /v1/projects/{projectId}/approvals/{approvalId}:
    get:
      tags: [Approvals]
      operationId: getApproval
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/ApprovalId"
      responses:
        "200":
          description: Current approval representation and strong aggregate-version validator.
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApprovalView"
        "401":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
  /v1/projects/{projectId}/approvals/{approvalId}/decision:
    post:
      tags: [Approvals]
      operationId: decideApproval
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/ApprovalId"
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/CorrelationId"
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovalDecisionRequest"
      responses:
        "200":
          description: Approval terminal decision, including an idempotently returned prior result.
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApprovalView"
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
        "412":
          $ref: "#/components/responses/Problem"
        "422":
          $ref: "#/components/responses/Problem"
...
components:
  securitySchemes:
    LocalSupervisorSession:
      type: apiKey
      in: cookie
      name: moonshift_supervisor_session
      description: >-
        Opaque host-only HttpOnly SameSite=Strict loopback evaluation session. It is never a provider
        credential and must not be exposed to JavaScript or logs.
  parameters:
    ProjectId:
      name: projectId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    ApprovalId:
      name: approvalId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Stable semantic command identity; 16 to 128 visible ASCII characters.
      schema:
        type: string
        minLength: 16
        maxLength: 128
        pattern: "^[!-~]+$"
    CorrelationId:
      name: X-Correlation-ID
      in: header
      required: true
      schema:
        type: string
        format: uuid
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: Strong ETag containing the aggregate version returned by the last query.
      schema:
        type: string
        pattern: '^"[1-9][0-9]*"$'
  headers:
    ETag:
      description: Strong aggregate-version ETag.
      schema:
        type: string
        pattern: '^"[1-9][0-9]*"$'
...
  responses:
...
    Problem:
      description: Safe problem detail with no stack, credential, or private reasoning.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
...
    ProjectState:
      type: string
      enum:
        - CREATING
        - ACTIVE
        - PAUSING
        - PAUSED
        - RESUMING
        - STOPPING
        - STOPPED
        - CANCELLING
        - COMPLETED
        - BLOCKED
        - FAILED
        - CANCELLED
...
    ExecutionState:
      type: string
      enum:
        - QUEUED
        - STARTING
        - RUNNING
        - WAITING_FOR_APPROVAL
        - CHECKPOINTING
        - SUSPENDED
        - STOPPING
        - STOPPED
        - SUCCEEDED
        - FAILED
        - CANCELLED
        - LOST
        - RECONCILING
...
    ApprovalState:
      type: string
      enum: [REQUESTED, APPROVED, REJECTED, EXPIRED, CANCELLED]
    ApprovalView:
      type: object
      additionalProperties: false
      required: [approvalId, projectId, taskId, requesterAgentId, state, actionDigest, scope, reason, riskSummary, expiresAt, version]
      properties:
        approvalId:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        taskId:
          type: string
          format: uuid
        requesterAgentId:
          type: string
          format: uuid
        state:
          $ref: "#/components/schemas/ApprovalState"
        actionDigest:
          type: string
          pattern: "^sha256:[a-f0-9]{64}$"
        scope:
          type: string
        reason:
          type: string
        riskSummary:
          type: string
        expiresAt:
          type: string
          format: date-time
        decidedAt:
          oneOf:
            - type: string
              format: date-time
            - type: "null"
        decisionActorId:
          oneOf:
            - type: string
              format: uuid
            - type: "null"
        version:
          type: integer
          minimum: 1
    ApprovalDecisionRequest:
      type: object
      additionalProperties: false
      required: [decision, actionDigest, reason]
      properties:
        decision:
          type: string
          enum: [APPROVE, REJECT]
        actionDigest:
          type: string
          pattern: "^sha256:[a-f0-9]{64}$"
        reason:
          type: string
          minLength: 1
          maxLength: 1000
    SupervisionView:
      type: object
      additionalProperties: false
      required: [items, budget, authority, effects, blockedReasons, projectState, projectVersion]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ApprovalView"
        budget:
          $ref: "#/components/schemas/SupervisionBudgetView"
        authority:
          $ref: "#/components/schemas/SupervisionAuthorityView"
        effects:
          type: array
          items:
            $ref: "#/components/schemas/EffectView"
        blockedReasons:
          type: array
          items:
            type: string
        projectState:
          $ref: "#/components/schemas/ProjectState"
        projectVersion:
          type: integer
          minimum: 1
    SupervisionBudgetView:
      type: object
      additionalProperties: false
      required: [invocationLimit, consumedInvocations, monetaryLimitMicros, consumedMonetaryMicros]
      properties:
        invocationLimit:
          type: integer
          minimum: 0
        consumedInvocations:
          type: integer
          minimum: 0
        monetaryLimitMicros:
          type: integer
          minimum: 0
        consumedMonetaryMicros:
          type: integer
          minimum: 0
    SupervisionAuthorityView:
      type: object
      additionalProperties: false
      required: [executionId, executionAttempt, executionState, capabilityLeaseId, capabilityLeaseState, capabilityLeaseExpiresAt, runnerLeaseId, runnerLeaseState, runnerLeaseExpiresAt, fencingToken, successor]
      properties:
        executionId:
          type: string
          format: uuid
        executionAttempt:
          type: integer
          minimum: 1
        executionState:
          $ref: "#/components/schemas/ExecutionState"
        capabilityLeaseId:
          type: string
          format: uuid
        capabilityLeaseState:
          type: string
          enum: [ACTIVE, SUSPENDED, REVOKED]
        capabilityLeaseExpiresAt:
          type: string
          format: date-time
        runnerLeaseId:
          type: string
          format: uuid
        runnerLeaseState:
          type: string
          enum: [ACTIVE, REVOKED]
        runnerLeaseExpiresAt:
          type: string
          format: date-time
        fencingToken:
          type: integer
          minimum: 1
          maximum: 9007199254740991
        successor:
          type: boolean
...
    EffectView:
      type: object
      additionalProperties: false
      required: [effectId, taskId, actionDigest, semanticKey, state, reconciliationOutcome, groundTruthDigest, version]
      properties:
        effectId:
          type: string
          format: uuid
        taskId:
          type: string
          format: uuid
        actionDigest:
          type: string
          pattern: "^sha256:[a-f0-9]{64}$"
        semanticKey:
          type: string
        state:
          type: string
          enum: [REQUESTED, EXECUTING, APPLIED, FAILED, UNKNOWN, RECONCILING, RECONCILED]
        reconciliationOutcome:
          oneOf:
            - type: string
            - type: "null"
        groundTruthDigest:
          oneOf:
            - type: string
              pattern: "^sha256:[a-f0-9]{64}$"
            - type: "null"
        version:
          type: integer
          minimum: 1
...
    Problem:
      type: object
      additionalProperties: false
      required: [type, title, status, code, correlationId]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        code:
          type: string
          pattern: "^[A-Z][A-Z0-9_]+$"
        detail:
          type: string
          maxLength: 2000
        correlationId:
          type: string
          format: uuid